typescript: transition test/service test/sync
