+++ test/sync/fetcher/headerfetcher.ts
@@ -1,6 +1,9 @@
-const tape = require('tape-catch')
+// Suppresses "Cannot redeclare block-scoped variable" errors
+// TODO: remove when import becomes possible
+export = {}
+
+import * as tape from 'tape-catch'
 const td = require('testdouble')
-// const BN = require('bn.js')
 const EventEmitter = require('events')
 const { defaultLogger } = require('../../../lib/logging')
 defaultLogger.silent = true
